3D point cloud reconstruction
The code subscribes to ros2 pointcloud topic, preprocesses the data, and then finds the centroid of the closest object in the scene.
After finding the closest object's centroid, goal points are generated around the object.
The robot goes to these goal points and creates pcd files. These files are then merged to create a one file reconstruction for the object.